+++ src/util/misc.h
@@ -25,7 +25,7 @@ string percent(
 
 void printVectorCSV(
   const string& title,
-  const vector<double>& peaks,
+  const vector<float>& peaks,
   const int level);
 
 #endif
